+++ Source/NonVisuals/Radios/RadioPanelPZ69Generic.cs
@@ -412,7 +412,7 @@ namespace NonVisuals.Radios
                                     PluginManager.DoEvent(
                                         DCSFPProfile.SelectedProfile.Description,
                                         HIDInstance,
-                                        (int)PluginGamingPanelEnum.PZ69RadioPanel,
+                                        PluginGamingPanelEnum.PZ69RadioPanel,
                                         (int)radioPanelKey.RadioPanelPZ69Knob,
                                         radioPanelKey.IsOn,
                                         keyBinding.OSKeyPress.KeyPressSequence);
@@ -430,7 +430,7 @@ namespace NonVisuals.Radios
                             PluginManager.DoEvent(
                                 DCSFPProfile.SelectedProfile.Description,
                                 HIDInstance,
-                                (int)PluginGamingPanelEnum.PZ69RadioPanel,
+                                PluginGamingPanelEnum.PZ69RadioPanel,
                                 (int)radioPanelKey.RadioPanelPZ69Knob,
                                 radioPanelKey.IsOn,
                                 null);
